Break benchmark task type cycle
