+++ src/gemmaclaw/benchmark/agent-tasks.ts
@@ -12,6 +12,12 @@ import {
   BENCHMARK_TEST_TEMPLATE_TARGETS,
   GENERATED_AGENT_VARIATION_TASKS,
 } from "./agent-variation-templates.js";
+import type {
+  AgentBenchmarkTask,
+  AgentDeterministicGrading,
+  AgentDeterministicScore,
+  DeterministicExpectedValue,
+} from "./agent-task-types.js";
 import { EXPANDED_AGENT_BENCHMARK_TASKS } from "./expanded-agent-benchmark-tasks.js";
 
 export {
@@ -19,76 +25,15 @@ export {
   EXPANDED_AGENT_BENCHMARK_TASKS,
   GENERATED_AGENT_VARIATION_TASKS,
 };
-
-export type AgentGradingType = "conversation_check" | "artifact_check" | "tool_sequence_check";
-
-export type AgentTaskCategory =
-  | "email"
-  | "calendar"
-  | "task_management"
-  | "multi_step"
-  | "security"
-  | "error_recovery"
-  | "memory"
-  | "ambiguous"
-  | "data_analysis"
-  | "coordination"
-  | "structured_output"
-  | "tool_intent"
-  | `expanded_${string}`
-  | `variant_${string}`;
-
-export type AgentTaskDifficulty = "easy" | "medium" | "hard" | "very_hard";
-
-export type DeterministicExpectedValue = string | string[];
-
-export type AgentDeterministicGrading =
-  | {
-      type: "json_fields";
-      requiredKeys: string[];
-      expectedFields: Record<string, DeterministicExpectedValue>;
-      allowExtraKeys?: boolean;
-    }
-  | {
-      type: "tool_intent";
-      allowedActions: string[];
-      expectedAction: string;
-      expectedArguments: Record<string, DeterministicExpectedValue>;
-      allowExtraTopLevelKeys?: boolean;
-      allowExtraArgumentKeys?: boolean;
-    };
-
-export type AgentDeterministicScore = {
-  taskId: string;
-  score: number;
-  maxScore: number;
-  percentage: number;
-  passed: boolean;
-  method: "deterministic";
-  details: string;
-};
-
-export type AgentBenchmarkTask = {
-  id: string;
-  name: string;
-  description: string;
-  category: AgentTaskCategory;
-  difficulty: AgentTaskDifficulty;
-  /** The prompt sent to the agent. */
-  prompt: string;
-  grading: {
-    type: AgentGradingType;
-    /** What the LLM judge checks for in the full conversation. */
-    criteria: string[];
-    /** Deterministic scorer used for small schema/intent tasks. */
-    deterministic?: AgentDeterministicGrading;
-    maxScore: number;
-  };
-  /** Deterministic mock response used by --mock mode when present. */
-  mock?: {
-    finalResponse: string;
-  };
-};
+export type {
+  AgentBenchmarkTask,
+  AgentDeterministicGrading,
+  AgentDeterministicScore,
+  AgentGradingType,
+  AgentTaskCategory,
+  AgentTaskDifficulty,
+  DeterministicExpectedValue,
+} from "./agent-task-types.js";
 
 export const OPENCLAW_HARD_WORKFLOW_TASK_IDS = [
   "recurring_template_persistence",
